requirements.txt
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.2.2
elasticsearch==8.14.0
python-dotenv==1.0.1
pydantic==2.8.2
fastapi==0.112.2
uvicorn==0.30.6
requests==2.32.3
streamlit==1.38.0